requirements-dev.txt
autopep8
category_encoders
cerberus
codecov
coverage
cpyquickhelper
cxxfilt
flake8
h5py
joblib
jupyter_sphinx
jyquickhelper
lightgbm
loky
matplotlib
mlinsights
mlprodict>=0.7
nbsphinx
onnxconverter-common
onnxmltools
openpyxl
pillow
py-spy
pandas
pandas_streaming>=0.3
pydata-sphinx-theme
pydot
pyinstrument
pylint
pyquickhelper>=1.10
pytest
pytest-cov
scikit-learn>=1.0
skl2onnx>=1.10.1
sphinx
sphinxcontrib-blockdiag
sphinxcontrib.imagesvg
sphinx-gallery
sympy
tabulate
torch
torchvision
tqdm
xgboost
wheel
